Promote 6 eligible TestE2EDynamicManifests subtests into TestE2EParallel's pool
The StorageClass/RWOP-conflict/VolumeAttachment/VolumeSnapshot, Crossplane XR, and
Helm release Secret subtests each already create and clean up their own dedicated
namespace and never touch cluster-wide state, matching TestE2EParallel's eligibility
bar -- they just hadn't been evaluated against it. Move them into new
runStorageSubtests/runCrossplaneSubtests/runHelmReleaseSubtests functions (following
the existing runXSubtests pattern) with t.Parallel(), so they join the concurrent pool
instead of running serially. TestE2EDynamicManifests now holds only the metrics-server,
VPA, RWOP-holder, PV-zone, WaitForFirstConsumer-topologies, and Karpenter-nodeSelector
subtests.

Fixes #832
